Move sab-proskomma to src/lib
We need to be using sab-proskomma in the PWA instead of just Proskomma
from proskomma-core so that we are using the updated selectors that
use BCP47 for the lang selector.
